Moved endpoint to PaymentsController also adjusted tests for it
